refactor components and screen integrations
