+++ thoughtbubble-mobile/src/components/navigation/ProjectsStackNavigator.tsx
@@ -1,11 +1,13 @@
 import React, { FC } from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
-import { ProjectsScreen } from '../ProjectsScreen';
-import { ThoughtsScreen } from '../ThoughtsScreen';
+import { ProjectsScreen } from '../../screens/ProjectsScreen';
+import { ThoughtsScreen } from '../../screens/ThoughtsScreen';
 import { ProjectStackParamList } from '../../interfaces/navigation';
 import { ProjectsStackNavigatorProps } from '../../interfaces/componentProps';
 import { colors } from '../../constants/colors';
 import { useDarkCheck } from '../../hooks/useDarkCheck';
+import { BlurView } from '@react-native-community/blur';
+import { View } from 'react-native';
 
 const { darkMode, lightMode } = colors;
 const ProjectStack = createStackNavigator<ProjectStackParamList>();
@@ -14,9 +16,22 @@ export const ProjectsStackNavigator: FC<ProjectsStackNavigatorProps> = () => {
   const isDarkMode = useDarkCheck();
 
   const dynamicScreenOptions = {
+    // blur 3/3
+    // headerTransparent: true,
+    // headerBackground: () => (
+    //   <BlurView
+    //     // style={{ flex: 1, backgroundColor: isDarkMode ? darkMode.background : lightMode.background }}
+    //     // style={{ flex: 1, backgroundColor: 'white' }}
+    //     style={{ flex: 1 }}
+    //     blurType={isDarkMode ? 'chromeMaterialDark' : 'light'}
+    //     blurAmount={15}
+    //     reducedTransparencyFallbackColor="black"
+    //   />
+    // ),
     headerStyle: {
       backgroundColor: isDarkMode ? darkMode.background : lightMode.background,
       // remove shadow working
+      // opacity: 0.5,
       elevation: 0,
       shadowOpacity: 0,
       borderBottomWidth: 0,
