+++ thoughtbubble-mobile/src/screens/SettingsScreen.tsx
