+++ thoughtbubble-mobile/src/screens/StatsHomeScreen.tsx
